containerize more components
