+++ common/docs/components/branchSelect.jsx
@@ -1,54 +1,27 @@
 import React from "react";
-import _ from "underscore";
 
-export default BranchSelect = React.createClass({
-  mixins: [ReactMeteorData],
-
-  getMeteorData() {
-    let branches = [];
-    let repo = ReDoc.Collections.Repos.findOne({ repo: this.props.repo }) || ReDoc.Collections.Repos.findOne();
-
-    if (repo && repo.branches) {
-      const isAdminUser = Roles.userIsInRole(Meteor.userId(), ["admin"], "redoc");
-
-      for (branch of repo.branches) {
-        if (Meteor.settings.public.redoc.publicBranches) {
-          const isPublicBranch = _.contains(Meteor.settings.public.redoc.publicBranches, branch.name);
-
-          if (isPublicBranch || isAdminUser) {
-            branches.push(branch.name);
-          }
-        } else {
-          branches.push(branch.name);
-        }
-      }
-    }
-
-    return {
-      branches: branches
-    };
-  },
+export default class BranchSelect extends React.Component {
 
   renderBranches() {
-    return this.data.branches.map((branch, index) => {
+    return this.props.branches.map((branch, index) => {
       return (
         <option key={index} value={branch}>{branch}</option>
       );
     });
-  },
+  }
 
   handleChange(event) {
     if (this.props.onBranchSelect) {
       this.props.onBranchSelect(event.target.value);
     }
-  },
+  }
 
   render() {
-    if (this.data.branches.length > 0) {
+    if (this.props.branches.length > 0) {
       return (
         <div className="redoc control select">
           <select
-            onChange={this.handleChange}
+            onChange={this.handleChange.bind(this)}
             ref="select"
             value={this.props.currentBranch}
             id="branch-select"
@@ -66,4 +39,4 @@ export default BranchSelect = React.createClass({
       <div className="loading">{"Loading..."}</div>
     );
   }
-});
+}
