+++ common/header/header.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import SearchField from "../search/search.jsx";
-import BranchSelect from "../docs/branchSelect.jsx";
+import BranchSelect from "../docs/containers/branchSelect";
 import { composeWithTracker } from 'react-komposer';
 import Avatar from "meteor/reactioncommerce:redoc-core/components/avatar.jsx"
